+++ sandbox/query_debugging/google_test.R
@@ -17,37 +17,7 @@ raw_results <- jsonlite::fromJSON(httr::content(soup, as = 'text', encoding = "U
 
 results <- tidygeocoder:::extract_results('google', raw_results)
 
-full_results <- tidygeocoder:::extract_results('google', raw_results, full_results = TRUE)
-
-########## test single values
-
-devtools::load_all()
-
-addr <- 'Tokyo, Japan'
-geo(address = addr, method = "osm", verbose = TRUE)
-geo(address = addr, method = "google", verbose = TRUE)
-geo_google(address = addr)
-
-# fails: google only takes single-line addresses
-geo(method = "google", 
-    verbose = TRUE,
-    street = "24 Sussex Drive",
-    city = "Ottawa",
-    country = "Canada")
-
-########## test columns in tibble
-
-addresses <- tibble::tibble(addrs = c("Tokyo, Japan", "Toronto, Ontario"))
-
-addresses %>%
-  geocode(addrs, method = 'google', verbose = TRUE)
-
-addresses %>%
-  geocode(addrs, method = 'google', verbose = TRUE, min_time = 1)
+full_results <- tidygeocoder:::extract_results('google', raw_results, full_results = TRUE, flatten = FALSE)
 
 
-addresses %>%
-  geocode(addrs, method = 'osm', verbose = TRUE)
 
-
-##################